Rename 'Endpoint' slice to 'Endpoints'; Add unit tests for DI extension method
